revisiones extras

+++ clases.modelo/ArticuloPrecio.cs
@@ -16,7 +16,7 @@ using System.Threading.Tasks;
         public decimal? Precio1 { get; set; }
         public decimal? Precio2 { get; set; }
         public decimal? Precio3 { get; set; }
-        public decimal? Relleno { get; set; }
+        public int? Relleno { get; set; }
         
 
     }
